pyproject.toml
[project]
name = "semantic-code-compression"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "matplotlib>=3.10.6",
    "nltk>=3.9.1",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.0",
    "tiktoken>=0.11.0",
    "torch>=2.8.0",
]
